Move rand DOI generator to util file
